# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IUserData": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"userFirstName": {
						"type": "string"
					},
					"userLastName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"userFirstName",
					"userLastName",
					"phone",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUserData.Exclude_keyofIUserData.userId__": {
				"properties": {
					"userFirstName": {
						"type": "string"
					},
					"userLastName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userFirstName",
					"userLastName",
					"phone",
					"email",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUserData.userId_": {
				"$ref": "#/components/schemas/Pick_IUserData.Exclude_keyofIUserData.userId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Partial_IUserData_": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"userFirstName": {
						"type": "string"
					},
					"userLastName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Pick_IUserData.Exclude_keyofIUserData.password__": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"userFirstName": {
						"type": "string"
					},
					"userLastName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"userFirstName",
					"userLastName",
					"phone",
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUserData.password_": {
				"$ref": "#/components/schemas/Pick_IUserData.Exclude_keyofIUserData.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
			"ILocationData": {
				"properties": {
					"locationId": {
						"type": "string"
					},
					"locationName": {
						"type": "string"
					},
					"locationAddress": {
						"type": "string"
					},
					"locationCity": {
						"type": "string"
					}
				},
				"required": [
					"locationId",
					"locationName",
					"locationAddress",
					"locationCity"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "care-idosos",
		"version": "1.0.0",
		"description": "Documentação da API Care Idosos",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Beatriz Ribeiro"
		}
	},
	"paths": {
		"/user/sign-up": {
			"post": {
				"operationId": "AddUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserData"
								}
							}
						}
					}
				},
				"tags": [
					"Acesso de Usuário"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IUserData.userId_"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Acesso de Usuário"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IUserData_"
							}
						}
					}
				}
			}
		},
...
		"/user/{userId}": {
			"get": {
				"operationId": "GetLocationById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Omit_IUserData.password_"
								}
							}
						}
					}
				},
				"tags": [
					"Acesso de Usuário"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/location/{locationId}": {
			"get": {
				"operationId": "GetLocationById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ILocationData"
								}
							}
						}
					}
				},
				"tags": [
					"Endereços das Clínicas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "locationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}